Split out controlled selection tests
