+++ package.json
@@ -45,7 +45,7 @@
   },
   "devDependencies": {
     "@eslint/js": "9.39.1",
-    "@storybook/react-vite": "10.1.0",
+    "@storybook/react-vite": "10.1.2",
     "@testing-library/react": "16.3.0",
     "@testing-library/user-event": "14.6.1",
     "@types/node": "24.10.1",
@@ -56,12 +56,12 @@
     "eslint": "9.39.1",
     "eslint-plugin-react": "7.37.5",
     "eslint-plugin-react-hooks": "7.0.1",
-    "eslint-plugin-storybook": "10.1.0",
+    "eslint-plugin-storybook": "10.1.2",
     "globals": "16.5.0",
     "jsdom": "27.2.0",
     "react": "19.2.0",
     "react-dom": "19.2.0",
-    "storybook": "10.1.0",
+    "storybook": "10.1.2",
     "typescript": "5.9.3",
     "typescript-eslint": "8.48.0",
     "vite": "7.2.4",
