+++ src/components/HighTable/HighTable.keyboard.test.tsx
@@ -1,8 +1,8 @@
-import { UserEvent } from '@testing-library/user-event'
+import type { UserEvent } from '@testing-library/user-event'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { createGetRowNumber, validateGetCellParams } from '../../helpers/dataframe/helpers.js'
-import { DataFrame } from '../../helpers/dataframe/index.js'
-import { OrderBy } from '../../helpers/sort.js'
+import type { DataFrame } from '../../helpers/dataframe/index.js'
+import type { OrderBy } from '../../helpers/sort.js'
 import { render } from '../../utils/userEvent.js'
 import HighTable from './HighTable.js'
 import type { Obj } from '../../helpers/dataframe/types.js'
